doc: 更新示例
