+++ src/BootstrapBlazor.Shared/Samples/Table/TablesDynamicObject.razor
@@ -1,24 +1,31 @@
-﻿@page "/DynamicTableSample"
+﻿@page "/tables/dynamicobject"
 
-<h3>TableColumn Template</h3>
+<h3>Table 表格</h3>
 
-<Table TItem="NetDynamic" Items="netDynamics" ShowToolbar="true" ShowColumnList="true" ShowDefaultButtons="false" ShowRefresh="false">
-    <TableColumns>
-        <TableColumn @bind-Field="@context.Fix" Sortable="true" Filterable="true" />
-        @foreach (var element in DynamicList)
-        {
-            <TableColumn @bind-Field="@context.Dynamic" Text="@element" Sortable="true" Filterable="true">
-                <Template Context="value">
-                    <div>@(value.Value.TryGetValue(element, out string? v) ? v : "")</div>
-                </Template>
-            </TableColumn>
-        }
-    </TableColumns>
-</Table>
+<h4>使用 <code>DynamicObject</code> 类作为数据源</h4>
+
+<DemoBlock Title="绑定集合" Introduction="通过 <code>OnQueryAsync</code> 回调获得数据集合" Name="OnQuery">
+    <Table TItem="CustomDynamicData" Items="DynamicItems"
+           IsStriped="true" IsBordered="true" ShowToolbar="true" ShowColumnList="true" ShowDefaultButtons="false" ShowRefresh="false">
+        <TableColumns>
+            <TableColumn @bind-Field="@context.Fix" Sortable="true" Filterable="true" />
+            @foreach (var element in DynamicList)
+            {
+                <TableColumn @bind-Field="@context.Dynamic" Text="@element" Sortable="true" Filterable="true">
+                    <Template Context="value">
+                        <div>@(value.Value.TryGetValue(element, out string? v) ? v : "")</div>
+                    </Template>
+                </TableColumn>
+            }
+        </TableColumns>
+    </Table>
+</DemoBlock>
+
 
 <h3>Net Dynamic</h3>
 
-<Table TItem="NetDynamic" Items="netDynamics" ShowToolbar="true" ShowColumnList="true" ShowDefaultButtons="false" ShowRefresh="false">
+<Table TItem="CustomDynamicData" Items="DynamicItems"
+       IsStriped="true" IsBordered="true" ShowToolbar="true" ShowColumnList="true" ShowDefaultButtons="false" ShowRefresh="false">
     <TableColumns>
         <TableColumn @bind-Field="@context.Fix" Sortable="true" Filterable="true" />
         @foreach (var element in DynamicList)
@@ -30,7 +37,8 @@
 
 <h3>BootstrapBlazor Dynamic</h3>
 
-<Table TItem="BBDynamic" Items="bbDynamics" ShowToolbar="true" ShowColumnList="true" ShowDefaultButtons="false" ShowRefresh="false">
+<Table TItem="BootstrapBlazorDynamicObjectData" Items="BootstrapDynamicItems"
+       IsStriped="true" IsBordered="true" ShowToolbar="true" ShowColumnList="true" ShowDefaultButtons="false" ShowRefresh="false">
     <TableColumns>
         <TableColumn @bind-Field="@context.Fix" Sortable="true" Filterable="true" />
         @foreach (var element in DynamicList)
